Split constraint package into api and impl
