+++ src/main/java/me/zeroeightsix/fiber/api/builder/constraint/AggregateConstraintsBuilder.java
@@ -1,8 +1,8 @@
 package me.zeroeightsix.fiber.api.builder.constraint;
 
 import me.zeroeightsix.fiber.api.builder.ConfigAggregateBuilder;
-import me.zeroeightsix.fiber.constraint.CompositeType;
-import me.zeroeightsix.fiber.constraint.Constraint;
+import me.zeroeightsix.fiber.api.constraint.CompositeType;
+import me.zeroeightsix.fiber.api.constraint.Constraint;
 import me.zeroeightsix.fiber.exception.RuntimeFiberException;
 
 import javax.annotation.Nonnull;
