+++ src/main/java/me/zeroeightsix/fiber/api/builder/constraint/ComponentConstraintsBuilder.java
@@ -1,9 +1,9 @@
 package me.zeroeightsix.fiber.api.builder.constraint;
 
 import me.zeroeightsix.fiber.api.builder.ConfigAggregateBuilder;
-import me.zeroeightsix.fiber.constraint.Constraint;
-import me.zeroeightsix.fiber.constraint.ConstraintType;
-import me.zeroeightsix.fiber.constraint.ValuedConstraint;
+import me.zeroeightsix.fiber.api.constraint.Constraint;
+import me.zeroeightsix.fiber.api.constraint.ConstraintType;
+import me.zeroeightsix.fiber.impl.constraint.ValuedConstraint;
 import me.zeroeightsix.fiber.exception.RuntimeFiberException;
 
 import javax.annotation.Nonnull;
