+++ src/main/java/me/zeroeightsix/fiber/api/constraint/Constraint.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.constraint;
+package me.zeroeightsix.fiber.api.constraint;
 
 import me.zeroeightsix.fiber.tree.ConfigLeaf;
