+++ src/main/java/me/zeroeightsix/fiber/impl/constraint/NumberConstraint.java
@@ -1,5 +1,6 @@
-package me.zeroeightsix.fiber.constraint;
+package me.zeroeightsix.fiber.impl.constraint;
 
+import me.zeroeightsix.fiber.api.constraint.ConstraintType;
 import me.zeroeightsix.fiber.exception.RuntimeFiberException;
 
 import java.math.BigDecimal;
