+++ src/main/java/me/zeroeightsix/fiber/impl/constraint/ValuedConstraint.java
@@ -1,4 +1,7 @@
-package me.zeroeightsix.fiber.constraint;
+package me.zeroeightsix.fiber.impl.constraint;
+
+import me.zeroeightsix.fiber.api.constraint.Constraint;
+import me.zeroeightsix.fiber.api.constraint.ConstraintType;
 
 /**
  * A constraint that has a value attached to it.
