+++ src/main/java/me/zeroeightsix/fiber/schema/SchemaGenerator.java
@@ -7,8 +7,8 @@ import blue.endless.jankson.JsonPrimitive;
 import blue.endless.jankson.impl.MarshallerImpl;
 import me.zeroeightsix.fiber.Identifier;
 import me.zeroeightsix.fiber.api.builder.constraint.CompositeConstraintsBuilder;
-import me.zeroeightsix.fiber.constraint.Constraint;
-import me.zeroeightsix.fiber.constraint.ValuedConstraint;
+import me.zeroeightsix.fiber.api.constraint.Constraint;
+import me.zeroeightsix.fiber.impl.constraint.ValuedConstraint;
 import me.zeroeightsix.fiber.serialization.Marshaller;
 import me.zeroeightsix.fiber.tree.ConfigTree;
 import me.zeroeightsix.fiber.tree.ConfigLeaf;
