+++ src/main/java/me/zeroeightsix/fiber/tree/ConfigLeafImpl.java
@@ -1,9 +1,9 @@
 package me.zeroeightsix.fiber.tree;
 
 import me.zeroeightsix.fiber.api.builder.ConfigLeafBuilder;
-import me.zeroeightsix.fiber.constraint.Constraint;
-import me.zeroeightsix.fiber.constraint.ConstraintType;
-import me.zeroeightsix.fiber.constraint.FinalConstraint;
+import me.zeroeightsix.fiber.api.constraint.Constraint;
+import me.zeroeightsix.fiber.api.constraint.ConstraintType;
+import me.zeroeightsix.fiber.impl.constraint.FinalConstraint;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
