+++ src/test/java/me/zeroeightsix/fiber/api/builder/constraint/ConstraintsBuilderTest.java
@@ -2,8 +2,8 @@ package me.zeroeightsix.fiber.api.builder.constraint;
 
 import me.zeroeightsix.fiber.api.builder.ConfigAggregateBuilder;
 import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
-import me.zeroeightsix.fiber.constraint.CompositeType;
-import me.zeroeightsix.fiber.constraint.Constraint;
+import me.zeroeightsix.fiber.api.constraint.CompositeType;
+import me.zeroeightsix.fiber.api.constraint.Constraint;
 import me.zeroeightsix.fiber.exception.RuntimeFiberException;
 import me.zeroeightsix.fiber.tree.ConfigLeaf;
 import org.junit.jupiter.api.DisplayName;
